[Foundation] Share custom constructor code between iOS and macOS for NSAttributedString.
This also makes the set of constructors for NSAttributedString identical
between all platforms.
