+++ src/Foundation/NSAttributedString.iOS.cs
@@ -39,23 +39,6 @@ using UIKit;
 
 namespace Foundation {
 
-#if !MONOMAC && !COREBUILD
-	public partial class NSAttributedString {
-		public NSAttributedString (NSUrl url, NSAttributedStringDocumentAttributes documentAttributes, ref NSError error)
-		: this (url, documentAttributes, out _, ref error) { }
-
-		public NSAttributedString (NSData data, NSAttributedStringDocumentAttributes documentAttributes, ref NSError error)
-		: this (data, documentAttributes, out _, ref error) { }
-
-		public NSAttributedString (NSUrl url, ref NSError error)
-		: this (url, new NSDictionary (), out _, ref error) { }
-
-		public NSAttributedString (NSData data, ref NSError error)
-		: this (data, new NSDictionary (), out _, ref error) { }
-
-	}
-#endif
-
 	public partial class NSAttributedStringDocumentAttributes : DictionaryContainer {
 #if !MONOMAC && !COREBUILD
 		public NSAttributedStringDocumentAttributes () : base (new NSMutableDictionary ()) { }
